1.8

+++ cryptocurrency-widgets-for-elementor/admin/settings.php
@@ -1,4 +1,9 @@
 <?php
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
 // Do not use namespace to keep this on global space to keep the singleton initialization working
 if (!class_exists('CCWE_Settings')) {
 
@@ -8,6 +13,8 @@ if (!class_exists('CCWE_Settings')) {
      *
      * Do not call or initialize this class directly, instead use the function mentioned at the bottom of this file
      */
+
+// phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedClassFound
     class CCWE_Settings
     {
 
@@ -25,7 +32,6 @@ if (!class_exists('CCWE_Settings')) {
                 return self::$instance = new self();
             }
             return self::$instance;
-
         }
 
         /**
@@ -38,14 +44,12 @@ if (!class_exists('CCWE_Settings')) {
 
             add_action('admin_enqueue_scripts', array($this, 'ccew_admin_style'));
             add_action('wp_ajax_ccew_delete_transient', array($this, 'ccew_delete_transient'));
-
         }
         public function ccew_admin_style()
         {
-            wp_enqueue_script('ccew-admin-script', CCEW_URL . 'assets/js/admin-script.js', array('jquery'), CCEW_VERSION);
+            wp_enqueue_script('ccew-admin-script', CCEW_URL . 'assets/js/admin-script.js', array('jquery'), CCEW_VERSION, true);
 
             wp_enqueue_style('ccew_admin_menu_style', CCEW_URL . 'assets/css/ccew-admin.css', null, CCEW_VERSION);
-
         }
 
         /**
@@ -56,14 +60,14 @@ if (!class_exists('CCWE_Settings')) {
             add_submenu_page('cool-crypto-plugins', 'Crypto Elementor Widget Settings', '<strong>Elementor Widget</strong>', 'manage_options', 'ccew-settings', array($this, 'Settings_callback'), 60);
 
             add_submenu_page('cool-crypto-plugins', 'Crypto Elementor Widget Settings', '↳ Settings', 'manage_options', 'ccew-settings', array($this, 'Settings_callback'), 61);
-
         }
 
         public function ccew_reg_settings()
         {
             register_setting(
                 'ccew_option_group', // Option group
-                'ccew-api-settings' // Option name
+                'ccew-api-settings', // Option name
+                'ccew_sanitize_api_settings' // Sanitize
             );
 
             add_settings_section(
@@ -76,7 +80,6 @@ if (!class_exists('CCWE_Settings')) {
             add_settings_field('Api_key', 'Select API', array($this, 'api_key_fun'), 'ccew-api-settings', 'ccew_section_id');
 
             add_settings_field('ccew_delete_cache', 'Purge API Data Cache', array($this, 'ccew_purge_cache'), 'ccew-api-settings', 'ccew_section_id');
-
         }
 
         /**
@@ -94,8 +97,7 @@ if (!class_exists('CCWE_Settings')) {
 
             $ajax_url = admin_url('admin-ajax.php');
             $ccew_nonce = wp_create_nonce('ccew-nonce');
-            echo ' <button class="button button-secondary ccew-delete-transient" data-ccpw-nonce="' . esc_attr($ccew_nonce) . '" data-ajax-url="' . esc_url($ajax_url) . '" id="ccew-delete-transient">' . __('Purge Cache', 'ccpw') . '</button>';
-
+            echo ' <button class="button button-secondary ccew-delete-transient" data-ccpw-nonce="' . esc_attr($ccew_nonce) . '" data-ajax-url="' . esc_url($ajax_url) . '" id="ccew-delete-transient">' . esc_html__('Purge Cache', 'cryptocurrency-widgets-for-elementor') . '</button>';
         }
 
         public function api_key_fun()
@@ -110,43 +112,54 @@ if (!class_exists('CCWE_Settings')) {
 
         public function Settings_callback()
         {
-            ?>
+?>
 
-			<?php settings_errors();?>
-			<form method="post" action="options.php">
-			<?php settings_fields('ccew_option_group');?>
-			<?php do_settings_sections('ccew-api-settings');?>
-			<?php submit_button();?>
-			</form>
-			<?php
+            <?php settings_errors(); ?>
+            <form method="post" action="options.php">
+                <?php settings_fields('ccew_option_group'); ?>
+                <?php do_settings_sections('ccew-api-settings'); ?>
+                <?php submit_button(); ?>
+            </form>
+<?php
 
-            $get_setings = isset($_GET['page']) ? $_GET['page'] : '';
-            $update_setings = isset($_GET['settings-updated']) ? $_GET['settings-updated'] : '';
+            // $get_setings = isset($_GET['page']) ? $_GET['page'] : '';
+
+            $get_setings = (isset($_GET['_wpnonce']) && wp_verify_nonce(sanitize_text_field(wp_unslash($_GET['_wpnonce'])), 'ccew_settings_nonce'))
+                ? sanitize_key(wp_unslash($_GET['page'] ?? ''))
+                : '';
+
+            $update_setings = sanitize_text_field(wp_unslash($_GET['settings-updated'] ?? ''));
 
             if ($get_setings == 'ccew-settings' && $update_setings == 'true') {
                 $db = new ccew_database();
                 $db->truncate_table();
                 delete_transient('ccew-saved-coindata');
                 delete_option('ccew_data_save');
-
             }
-
         }
         public function ccew_delete_transient()
         {
             // Check for nonce security
-            if (!wp_verify_nonce($_POST['nonce'], 'ccew-nonce')) {
-                die('You don\'t have permission to delete the cache.');
+            // if (!wp_verify_nonce($_POST['nonce'], 'ccew-nonce')) {
+            //     die('You don\'t have permission to delete the cache.');
+            // }
+            if (
+                ! isset($_POST['nonce']) ||
+                ! wp_verify_nonce(
+                    sanitize_text_field(wp_unslash($_POST['nonce'])),
+                    'ccew-nonce'
+                )
+            ) {
+                wp_die(esc_html__('You don\'t have permission to delete the cache.', 'cryptocurrency-widgets-for-elementor'));
             }
+
             // Delete cache if user has permssion to delete it.
             if (current_user_can('manage_options')) {
                 delete_transient('ccew-saved-coindata');
                 delete_option('ccew_data_save');
                 wp_send_json_success();
             }
-
         }
-
     }
 
     /**
@@ -154,7 +167,7 @@ if (!class_exists('CCWE_Settings')) {
      * initialize the main dashboard class with all required parameters
      */
 
+    // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedVariableFound
     $Openexchange = CCWE_Settings::init();
     $Openexchange->ccwe_init_hooks();
-
 }
